Activity Selection Greedy Algorithm
The Activity Selection Problem is a classic example of a greedy algorithm. The goal is to select the maximum number of activities that don't overlap, given a set of activities with their start and finish times.

Steps in the Greedy Approach:
Sort the activities by their finish time in ascending order.
Select the first activity (the one with the earliest finish time) since it leaves the most time for the remaining activities.
For each subsequent activity, select it if its start time is greater than or equal to the finish time of the previously selected activity.
Repeat until all activities are processed.